aaaa

+++ src/main/java/com/example/backend/Services/CustomerService.java
@@ -1,17 +1,25 @@
 package com.example.backend.Services;
 
+import com.example.backend.Config.SecurityConfig;
 import com.example.backend.Entity.Customer;
 import com.example.backend.Repositories.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 @Service
-public class CustomerService {
+public class CustomerService implements UserDetailsService {
 
     @Autowired
     private CustomerRepository customerRepository;
+
+    @Autowired
+    BCryptPasswordEncoder bCryptPasswordEncoder;
 
     public ResponseEntity<?> getCustomer(Integer id){
         Customer customer = customerRepository.findById(id).orElse(null);
@@ -34,11 +42,27 @@ public class CustomerService {
             System.out.println("чувак уже есть!");
             return false;
         }
+        customer.setPassword(bCryptPasswordEncoder.encode(customer.getPassword()));
+
         customerRepository.save(newCustomer);
         return true;
     }
 
-    public void deleteCustomer(Integer id){
-        customerRepository.deleteById(id);
+    public boolean deleteCustomer(Integer id){
+        if (customerRepository.findById(id).isPresent()) {
+            customerRepository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        Customer customer = customerRepository.findByEmail(email);
+
+        if (customer == null) {
+            throw new UsernameNotFoundException("User not found");
+        }
+        return customer;
     }
 }
